Implement support for registering multiple martians with re-frame
By default if no custom id for the martian is supplied it is
registered under `::martian.reframe/default-id` and the parameters for
a request using this martian does not need to change.

If a martian registered with a custom instance id `init` should be
called as `(martian.re-frame/init "https://some.url" {::martian.re-frame/instance-id :custom-id})`
and then the parameters for a request shold contain a
`::martian.re-frame/instance-id` key referencing the id used at init.