Showcase that everything can work with `:auto`
Both `clj-http` and `hato` clients are capable of doing the Content-Type-based
response coercions (the `:as :auto` request option) when the built-in encoders
for the client-supported media types get skipped. Which effectively means that
we don't need these encoders for those HTTP clients which are capable of doing
the output coercion on their own; the built-in encoders are only necessary for
other barebone clients, such as `http-kit`, `clj-http-lite`, `bb-http-client`,
etc., that only coerce bodies to raw types (strings, streams, or bytes).